[Cpp] Wipe the include of <Windows.h> in Context.h and MemoryMappedFile.h
