+++ aeron-client/src/main/cpp/Context.cpp
@@ -19,6 +19,13 @@
 #include "Aeron.h"
 #include "CncFileDescriptor.h"
 
+#if defined (_WIN32)
+    #ifndef NOMINMAX
+        #define NOMINMAX
+    #endif // !NOMINMAX
+    #include <Windows.h>
+#endif
+
 using namespace aeron;
 using namespace aeron::util;
 
@@ -48,4 +55,62 @@ void Context::requestDriverTermination(
 
         driverProxy.terminateDriver(tokenBuffer, tokenLength);
     }
+}
+
+#if !defined(__linux__)
+inline static std::string tmpDir()
+{
+#if defined(_MSC_VER)
+    static char buff[MAX_PATH+1];
+    std::string dir = "";
+
+    if (::GetTempPath(MAX_PATH, &buff[0]) > 0)
+    {
+        dir = buff;
+    }
+
+    return dir;
+#else
+    std::string dir = "/tmp";
+
+    if (::getenv("TMPDIR"))
+    {
+        dir = ::getenv("TMPDIR");
+    }
+
+    return dir;
+#endif
+}
+#endif
+
+inline static std::string getUserName()
+{
+    const char *username = ::getenv("USER");
+#if (_MSC_VER)
+    if (nullptr == username)
+    {
+        username = ::getenv("USERNAME");
+        if (nullptr == username)
+        {
+                username = "default";
+        }
+    }
+#else
+    if (nullptr == username)
+    {
+        username = "default";
+    }
+#endif
+    return username;
+}
+
+std::string Context::defaultAeronPath()
+{
+#if defined(__linux__)
+    return "/dev/shm/aeron-" + getUserName();
+#elif (_MSC_VER)
+    return tmpDir() + "aeron-" + getUserName();
+#else
+    return tmpDir() + "/aeron-" + getUserName();
+#endif
 }
